Created ProfileScreenViewModel
